Working and clean, performance is not optimal though because excessive callbacks and operations

+++ src/renderer/ui/components/ShowForm.tsx
@@ -4,7 +4,7 @@ import MomentUtils from "@date-io/moment"
 import moment, {Moment} from "moment/moment"
 import {Select, Typography} from "@material-ui/core"
 import MenuItem from "@material-ui/core/MenuItem"
-import {allDays, json} from "../../../shared/Utils"
+import {allDays} from "../../../shared/Utils"
 import {Day} from "../../../shared/Types"
 import TextField from "@material-ui/core/TextField"
 import Button from "@material-ui/core/Button"
@@ -49,7 +49,9 @@ export class ShowForm extends Component <ShowFormProps, ShowFormState> {
     constructor(props: ShowFormProps) {
         super(props)
         this.onChange = this.onChange.bind(this)
+        this.onNameChanged = this.onNameChanged.bind(this)
         this.onDayChanged = this.onDayChanged.bind(this)
+        this.onTimeChanged = this.onTimeChanged.bind(this)
         this.onDurationChanged = this.onDurationChanged.bind(this)
         this.state = SerializedShowToShowFormState(this.props.serializedShow.name, this.props.serializedShow)
         RequestSender.getAllSettings().then((allSettings: AllSettings) => {
@@ -59,30 +61,37 @@ export class ShowForm extends Component <ShowFormProps, ShowFormState> {
 
     public async onChange(): Promise<void> {
         const serializedShow = ShowFormStateToSerializedShow(this.state)
-        console.log(json(this.state))
         if (this.props.onChange) this.props.onChange(this.state.originalName, serializedShow)
+    }
+
+    public onNameChanged(nameString: string) {
+        this.setState({name: nameString}, this.onChange)
     }
 
     public onDayChanged(dayString: string) {
         const day: Day = dayString as Day
-        this.setState({day: day}, () => this.onChange())
+        this.setState({day: day}, this.onChange)
+    }
+
+    public onTimeChanged(date: Moment | null) {
+        if (date) {
+            this.setState({startTime: {h: date.hours(), m: date.minutes()}}, this.onChange)
+        }
     }
 
     public onDurationChanged(durationString: string) {
-        console.log(durationString)
         const duration: number = durationString === "" ? 0 : Number.parseInt(durationString)
-        this.setState({duration: {amount: duration, unit: "minutes"}}, () => {
-            this.onChange()
-        })
+        this.setState({duration: {amount: duration, unit: "minutes"}}, this.onChange)
     }
 
     public render() {
         return (
             <MuiPickersUtilsProvider utils={MomentUtils}>
                 <Typography style={{color: "black"}} variant="h5">{this.state.name}</Typography>
-                <TextField label="Show name"
+                <TextField autoFocus
+                           label="Show name"
                            value={this.state.name}
-                           onChange={(event) => this.setState({name: event.target.value})}/>
+                           onChange={(event) => this.onNameChanged(event.target.value)}/>
                 <Typography style={{color: "black"}} variant="h6">Day</Typography>
                 <Select value={this.state.day}
                         onChange={(event) => this.onDayChanged(event.target.value as string)}>
@@ -93,26 +102,14 @@ export class ShowForm extends Component <ShowFormProps, ShowFormState> {
                     ampm={false}
                     label="24 hours"
                     value={moment(this.state.startTime)}
-                    onChange={(date: Moment | null) => {
-                        if (date) {
-                            const obj = date.toObject()
-                            console.log(obj)
-                            const mom = moment(obj)
-                            const oobj = mom.toObject()
-                            console.log(oobj)
-                            console.log(obj == oobj)
-                            this.setState({startTime: {h: date.hours(), m: date.minutes()}})
-                        }
-                    }}
+                    onChange={this.onTimeChanged}
                 />
                 <Typography style={{color: "black"}} variant="h6">Duration</Typography>
                 <TextField label="Duration (minutes)" type="number"
                            value={this.state.duration.amount}
                            onChange={(event) => this.onDurationChanged(event.target.value)}/>
                 <Button onClick={() => {
-                    if (this.props.onDelete) {
-                        this.props.onDelete(this.state.originalName, this.state)
-                    }
+                    if (this.props.onDelete) this.props.onDelete(this.state.originalName, this.state)
                 }}><Delete/>Delete Show</Button>
             </MuiPickersUtilsProvider>
         )
